pyproject.toml
[project]
name = "mcp-api-mock-gen"
version = "0.1.0"
description = "MCP server for rapidly creating mocked REST APIs from JSON examples"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "github-copilot-sdk",
    "azure-cosmos>=4.9.0",
    "azure-identity>=1.19.0",
    "openai>=1.60.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_api_mock_gen"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
